+++ header.h
@@ -2,18 +2,19 @@
 #include <vector>
 #include <iostream>
 #include <random>
+#include <stdio.h>
 
 /* Structure to hold the current properties of the Ising Model */
 struct Ising_Model
 {
 	Ising_Model(int n_rows, int n_cols, double T, long N_steps);  // signature for constructor, construct instance of class
 
-	std::vector<int> states; // 1d array to hold the binary name of all states
+	std::vector<unsigned long long> states; // 1d array to hold the name of all states
 	std::vector<std::vector<int>> spin_matrix;	// 2d array for the current spin matrix
-	int current_step; // integer denoting the number of steps taken
+	long current_step; // integer denoting the number of steps taken
 	int num_rows; // number of rows
 	int num_cols; //number of columns
-	int num_steps; // number of simulation steps
+	long num_steps; // number of simulation steps
 	double temp; // temperature of the model
 
 	/* Define Methods */
@@ -25,4 +26,4 @@ struct Ising_Model
 std::ostream &operator<<(std::ostream &out, std::vector<std::vector<int>> spin_matrix);
 
 /* Other Functions */
-int get_state(std::vector<std::vector<int>> spin_matrix, int num_rows, int num_cols);
+unsigned long long get_state(std::vector<std::vector<int>> spin_matrix, int num_rows, int num_cols);
